Refactor agent request modules
